+++ src/ApiService/ApiService/QueueNodeHearbeat.cs
@@ -16,15 +16,20 @@ public class QueueNodeHearbeat
     private readonly ILogger _logger;
     private readonly IStorageProvider _storageProvider;
 
-    public QueueNodeHearbeat(ILoggerFactory loggerFactory, IStorageProvider storageProvider)
+    private readonly IEvents _events;
+
+    public QueueNodeHearbeat(ILoggerFactory loggerFactory, IStorageProvider storageProvider, IEvents events)
     {
         _logger = loggerFactory.CreateLogger<QueueNodeHearbeat>();
         _storageProvider = storageProvider;
+        _events = events;
     }
 
     [Function("QueueNodeHearbeat")]
     public async Task Run([QueueTrigger("myqueue-items", Connection = "AzureWebJobsStorage")] string msg)
     {
+        _logger.LogInformation($"heartbeat: {msg}");
+
         var hb = JsonSerializer.Deserialize<NodeHeartbeatEntry>(msg, EntityConverter.GetJsonSerializerOptions()).EnsureNotNull($"wrong data {msg}");
 
         var node = await Node.GetByMachineId(_storageProvider, hb.NodeId);
@@ -38,14 +43,6 @@ public class QueueNodeHearbeat
 
         await _storageProvider.Replace(newNode);
 
-        //send_event(
-        //    EventNodeHeartbeat(
-        //        machine_id = node.machine_id,
-        //        scaleset_id = node.scaleset_id,
-        //        pool_name = node.pool_name,
-        //    )
-        //)
-
-        _logger.LogInformation($"heartbeat: {msg}");
+        await _events.SendEvent(new EventNodeHeartbeat(node.MachineId, node.ScalesetId, node.PoolName));
     }
 }
