+++ src/ApiService/ApiService/onefuzzlib/orm/StorageProvider.cs
@@ -14,8 +14,8 @@ namespace Microsoft.OneFuzz.Service.OneFuzzLib.Orm;
 public interface IStorageProvider
 {
     Task<TableClient> GetTableClient(string table);
-    IAsyncEnumerable<T> QueryAsync<T>(string filter) where T : EntityBase;
-    Task<bool> Replace<T>(T entity) where T : EntityBase;
+    //IAsyncEnumerable<T> QueryAsync<T>(string filter) where T : EntityBase;
+    //Task<bool> Replace<T>(T entity) where T : EntityBase;
 
 }
 
@@ -48,21 +48,5 @@ public class StorageProvider : IStorageProvider
         return (resourceId.Name, key?.Value);
     }
 
-    public async IAsyncEnumerable<T> QueryAsync<T>(string filter) where T : EntityBase
-    {
-        var tableClient = await GetTableClient(typeof(T).Name);
-
-        await foreach (var x in tableClient.QueryAsync<TableEntity>(filter).Select(x => _entityConverter.ToRecord<T>(x))) {
-            yield return x;
-        }
-    }
-
-    public async Task<bool> Replace<T>(T entity) where T : EntityBase
-    {
-        var tableClient = await GetTableClient(typeof(T).Name);
-        var tableEntity = _entityConverter.ToTableEntity(entity);
-        var response = await tableClient.UpsertEntityAsync(tableEntity);
-        return !response.IsError;
-
-    }
+    
 }
